backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
httpx>=0.27.0
ffmpeg-python>=0.2.0
Pillow>=10.4.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
fonttools>=4.0.0
deepl>=1.18.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
slowapi>=0.1.9
litellm>=1.40.0
